Main: add argument to explicitly set Base IRI for generated terms
(R2)RML spec requires an executing environment variable to dictate the Base IRI, not the @base directive of the mapping file. The Main.main method now only uses the @base directive as base for generated terms when no explicit base IRI is set in the CLI, and strict mode is not enabled.