XMLRecord: fallback to String on NodeSet failure
By default, the RMLMapper evaluates an XPath expression and expects
an XML NODESET as result. However, in some cases this can be either
an BOOLEAN, STRING or NUMBER and therefore raises an exception.
If that's the case, try to evaluate the XPath expression again and
return the results as an Java String. If that still fails, a faulty
XPath expression is provided which will print an error to the user.

Fixes https://github.com/RMLio/rmlmapper-java/issues/128