XMLRecordFactory: add initial support for XML namespaces
XML namespaces should be provided to the XPath engine before parsing the
XML file. RML mapping rules do not specify these namespaces, therefore
we parse the XML file and search for the namespace definitions. These
namespaces are extracted and added to the XPath engine to execute the
XPath query in the RML mapping rules. Only fully namespaced XML files
are currently supported.