functions: lib: IDLabFunctions: functions lookup and lookupWithDelimiter
Implementation of a function that looks for the first occurence of a certain value in a column of a csv file, in order to return a value from a different column in the same row.
It will be used in the AI4Fl project to align ID's between different source systems.