target: TargetFactory: expand LDES EventStreamTarget for ActivityStreams
The original LDES EvenStreamTarget uses the IRIs of the generated
triples for LDES members. However, some vocabularies like W3C Activity
Streams puts these IRIs as an object of certain predicates like
as:object.

Add support to the LDES EvenStreamTarget to extract these IRIs by adding
an optional ldes:memberPath property. If this property is absent, the
original approach of using the subject IRIs of all triples is used.
Also make the member IRIs immutable if required by setting the
ldes:generateUniqueIRI property to 'true'. If absent or 'false', the
IRIs are extracted and used as such.