SolidTargetTest: use the host the Solid Docker container runs on instead of 'localhost'.
It is not guaranteed that the Docker container running a Solid server that is
started in SolidTargetTest runs on address 'localhost', for example in the GitLab
CI tests. Use the hostname provided by Docker instead.

As a consequence, the mapping files in src/test/resources/solid-target use a template
for the host which is then replaced with the Docker host at run time.